# estransport: Fix data race on subsequent iteration in retry loop