This fixes the data race of the http.Request.URL field as it is read in
another goroutine by http.Transport when it tries to write to the
persistent connection. The fix is to clone the request on subsequent
iteration.

Closes #147