@karmi 2020-08-08
Hello! Thanks for the patch! I'm out of office this month, so have a rather limited time to look into these things. Couple of comments:

* Can you please run the benchmarking suite (`test-bench`) and have a look if the request cloning has an adverse effect on memory allocation?

* I'm not sure that we should add a separate make target for this isolated case. Also, `docker start elasticsearch` is perhaps a custom command you have? The repository contains the `make cluster` target for launching Elasticsearch, so that's what should be used. I imagine that we can add a sequence of commands to the CI config, which would run more-or-less what your target does, without adding to the Makefile surface.