@pengux 2020-08-10
@karmi Here's the comparison for the benchmarks from `test-bench` using `benchcmp`:

```
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkClient/Create_client_with_defaults-12                   19134         19904         +4.02%
BenchmarkClientAPI/InfoRequest{}.Do()-12                         940           958           +1.91%
BenchmarkClientAPI/IndexRequest{...}.Do()-12                     2776          2764          -0.43%
BenchmarkClientAPI/Index(...)-12                                 3045          3123          +2.56%
BenchmarkClientAPI/SearchRequest{...}.Do()-12                    2641          2734          +3.52%
BenchmarkClientAPI/Search(...)-12                                3206          3212          +0.19%
BenchmarkClientAPI/BulkRequest{...}.Do()-12                      2953          2715          -8.06%
BenchmarkClientAPI/Bulk(...)-12                                  2971          3113          +4.78%
BenchmarkAPI/client.Info()______________________-12              868           874           +0.69%
BenchmarkAPI/client.Info()_WithContext__________-12              1135          1072          -5.55%
BenchmarkAPI/InfoRequest{}.Do()_________________-12              999           996           -0.30%
BenchmarkAPI/client.Cluster.Health()____________-12              1096          1008          -8.03%
BenchmarkAPI/client.Cluster.Health()_With()_____-12              2424          2340          -3.47%
BenchmarkAPI/ClusterHealthRequest{}.Do()________-12              1163          1180          +1.46%
BenchmarkAPI/ClusterHealthRequest{...}.Do()_____-12              2065          2049          -0.77%
BenchmarkAPI/client.Index()_With()_____________-12               3060          2928          -4.31%
BenchmarkAPI/IndexRequest{...}.Do()_____________-12              2767          2870          +3.72%
BenchmarkAPI/IndexRequest{...}.Do()_reused______-12              2660          2816          +5.86%
BenchmarkAPI/client.Search()___________________-12               1597          1574          -1.44%
BenchmarkAPI/client.Search()_with_error________-12               1776          1748          -1.58%
BenchmarkAPI/SearchRequest{...}.Do()____________-12              1286          1270          -1.24%
BenchmarkAPI/SearchRequest{...}.Do()_with_error_-12              1425          1419          -0.42%
BenchmarkSingleConnectionPool/Next()/Single__________-12         0.31          0.30          -3.87%
BenchmarkSingleConnectionPool/Next()/Parallel_(1000)-12          0.20          0.21          +2.97%
BenchmarkSingleConnectionPool/OnFailure()/Single_____-12         0.31          0.30          -2.57%
BenchmarkSingleConnectionPool/OnFailure()/Parallel_(1000)-12     0.19          0.20          +4.21%
BenchmarkStatusConnectionPool/Next()/Single_____-12              40.2          40.0          -0.50%
BenchmarkStatusConnectionPool/Next()/Parallel_(100)-12           194           206           +6.19%
BenchmarkStatusConnectionPool/Next()/Parallel_(1000)-12          197           157           -20.30%
BenchmarkStatusConnectionPool/OnFailure()/Single_____-12         29.0          28.5          -1.72%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(10)-12       110           112           +1.82%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(100)-12      117           121           +3.42%
BenchmarkStatusConnectionPool/OnSuccess()/Single_____-12         16.1          16.8          +4.35%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(10)-12       89.9          94.1          +4.67%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(100)-12      10.1          8.54          -15.45%
BenchmarkStatusConnectionPool/resurrect()/Single-12              66.7          66.2          -0.75%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(10)-12       14533         15955         +9.78%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(100)-12      16200         16145         -0.34%
BenchmarkTransport/Defaults-12                                   1407          1398          -0.64%
BenchmarkTransport/Headers-12                                    1589          1580          -0.57%
BenchmarkTransportLogger/Text-12                                 2627          2602          -0.95%
BenchmarkTransportLogger/Text-Body-12                            7219          7405          +2.58%
BenchmarkTransportLogger/JSON-12                                 2569          2670          +3.93%
BenchmarkTransportLogger/JSON-Body-12                            5160          5247          +1.69%
BenchmarkBulkIndexer/Basic-12                                    955           952           -0.31%
BenchmarkJSONReader/None-12                                      858           808           -5.83%
BenchmarkJSONReader/Default-12                                   1526          1508          -1.18%
BenchmarkJSONReader/Default-Copy-12                              1065          909           -14.65%
BenchmarkJSONReader/Custom-12                                    817           784           -4.04%

benchmark                                                        old allocs     new allocs     delta
BenchmarkClient/Create_client_with_defaults-12                   372            372            +0.00%
BenchmarkClientAPI/InfoRequest{}.Do()-12                         9              9              +0.00%
BenchmarkClientAPI/IndexRequest{...}.Do()-12                     27             27             +0.00%
BenchmarkClientAPI/Index(...)-12                                 33             33             +0.00%
BenchmarkClientAPI/SearchRequest{...}.Do()-12                    26             26             +0.00%
BenchmarkClientAPI/Search(...)-12                                34             34             +0.00%
BenchmarkClientAPI/BulkRequest{...}.Do()-12                      28             28             +0.00%
BenchmarkClientAPI/Bulk(...)-12                                  33             33             +0.00%
BenchmarkAPI/client.Info()______________________-12              8              8              +0.00%
BenchmarkAPI/client.Info()_WithContext__________-12              12             12             +0.00%
BenchmarkAPI/InfoRequest{}.Do()_________________-12              9              9              +0.00%
BenchmarkAPI/client.Cluster.Health()____________-12              8              8              +0.00%
BenchmarkAPI/client.Cluster.Health()_With()_____-12              22             22             +0.00%
BenchmarkAPI/ClusterHealthRequest{}.Do()________-12              9              9              +0.00%
BenchmarkAPI/ClusterHealthRequest{...}.Do()_____-12              18             18             +0.00%
BenchmarkAPI/client.Index()_With()_____________-12               31             31             +0.00%
BenchmarkAPI/IndexRequest{...}.Do()_____________-12              27             27             +0.00%
BenchmarkAPI/IndexRequest{...}.Do()_reused______-12              27             27             +0.00%
BenchmarkAPI/client.Search()___________________-12               15             15             +0.00%
BenchmarkAPI/client.Search()_with_error________-12               18             18             +0.00%
BenchmarkAPI/SearchRequest{...}.Do()____________-12              11             11             +0.00%
BenchmarkAPI/SearchRequest{...}.Do()_with_error_-12              14             14             +0.00%
BenchmarkSingleConnectionPool/Next()/Single__________-12         0              0              +0.00%
BenchmarkSingleConnectionPool/Next()/Parallel_(1000)-12          0              0              +0.00%
BenchmarkSingleConnectionPool/OnFailure()/Single_____-12         0              0              +0.00%
BenchmarkSingleConnectionPool/OnFailure()/Parallel_(1000)-12     0              0              +0.00%
BenchmarkStatusConnectionPool/Next()/Single_____-12              0              0              +0.00%
BenchmarkStatusConnectionPool/Next()/Parallel_(100)-12           0              0              +0.00%
BenchmarkStatusConnectionPool/Next()/Parallel_(1000)-12          0              0              +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Single_____-12         0              0              +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(10)-12       0              0              +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(100)-12      0              0              +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Single_____-12         0              0              +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(10)-12       0              0              +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(100)-12      0              0              +0.00%
BenchmarkStatusConnectionPool/resurrect()/Single-12              0              0              +0.00%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(10)-12       0              0              +0.00%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(100)-12      0              0              +0.00%
BenchmarkTransport/Defaults-12                                   12             12             +0.00%
BenchmarkTransport/Headers-12                                    13             13             +0.00%
BenchmarkTransportLogger/Text-12                                 21             21             +0.00%
BenchmarkTransportLogger/Text-Body-12                            40             40             +0.00%
BenchmarkTransportLogger/JSON-12                                 16             16             +0.00%
BenchmarkTransportLogger/JSON-Body-12                            28             28             +0.00%
BenchmarkBulkIndexer/Basic-12                                    3              3              +0.00%
BenchmarkJSONReader/None-12                                      8              8              +0.00%
BenchmarkJSONReader/Default-12                                   13             13             +0.00%
BenchmarkJSONReader/Default-Copy-12                              10             10             +0.00%
BenchmarkJSONReader/Custom-12                                    8              8              +0.00%

benchmark                                                        old bytes     new bytes     delta
BenchmarkClient/Create_client_with_defaults-12                   14632         14632         +0.00%
BenchmarkClientAPI/InfoRequest{}.Do()-12                         1217          1217          +0.00%
BenchmarkClientAPI/IndexRequest{...}.Do()-12                     2058          2058          +0.00%
BenchmarkClientAPI/Index(...)-12                                 2458          2457          -0.04%
BenchmarkClientAPI/SearchRequest{...}.Do()-12                    2000          2000          +0.00%
BenchmarkClientAPI/Search(...)-12                                2896          2896          +0.00%
BenchmarkClientAPI/BulkRequest{...}.Do()-12                      2312          2312          +0.00%
BenchmarkClientAPI/Bulk(...)-12                                  2680          2680          +0.00%
BenchmarkAPI/client.Info()______________________-12              897           897           +0.00%
BenchmarkAPI/client.Info()_WithContext__________-12              1321          1321          +0.00%
BenchmarkAPI/InfoRequest{}.Do()_________________-12              1217          1217          +0.00%
BenchmarkAPI/client.Cluster.Health()____________-12              1088          1088          +0.00%
BenchmarkAPI/client.Cluster.Health()_With()_____-12              2120          2120          +0.00%
BenchmarkAPI/ClusterHealthRequest{}.Do()________-12              1248          1248          +0.00%
BenchmarkAPI/ClusterHealthRequest{...}.Do()_____-12              1800          1800          +0.00%
BenchmarkAPI/client.Index()_With()_____________-12               2384          2384          +0.00%
BenchmarkAPI/IndexRequest{...}.Do()_____________-12              2057          2058          +0.05%
BenchmarkAPI/IndexRequest{...}.Do()_reused______-12              2058          2058          +0.00%
BenchmarkAPI/client.Search()___________________-12               2080          2080          +0.00%
BenchmarkAPI/client.Search()_with_error________-12               2272          2272          +0.00%
BenchmarkAPI/SearchRequest{...}.Do()____________-12              1296          1296          +0.00%
BenchmarkAPI/SearchRequest{...}.Do()_with_error_-12              1488          1488          +0.00%
BenchmarkSingleConnectionPool/Next()/Single__________-12         0             0             +0.00%
BenchmarkSingleConnectionPool/Next()/Parallel_(1000)-12          0             0             +0.00%
BenchmarkSingleConnectionPool/OnFailure()/Single_____-12         0             0             +0.00%
BenchmarkSingleConnectionPool/OnFailure()/Parallel_(1000)-12     0             0             +0.00%
BenchmarkStatusConnectionPool/Next()/Single_____-12              0             0             +0.00%
BenchmarkStatusConnectionPool/Next()/Parallel_(100)-12           0             0             +0.00%
BenchmarkStatusConnectionPool/Next()/Parallel_(1000)-12          0             0             +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Single_____-12         0             0             +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(10)-12       0             0             +0.00%
BenchmarkStatusConnectionPool/OnFailure()/Parallel_(100)-12      0             0             +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Single_____-12         0             0             +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(10)-12       0             0             +0.00%
BenchmarkStatusConnectionPool/OnSuccess()/Parallel_(100)-12      0             0             +0.00%
BenchmarkStatusConnectionPool/resurrect()/Single-12              38            39            +2.63%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(10)-12       0             0             +0.00%
BenchmarkStatusConnectionPool/resurrect()/Parallel_(100)-12      1             2             +100.00%
BenchmarkTransport/Defaults-12                                   1336          1336          +0.00%
BenchmarkTransport/Headers-12                                    1352          1352          +0.00%
BenchmarkTransportLogger/Text-12                                 1624          1624          +0.00%
BenchmarkTransportLogger/Text-Body-12                            12248         12248         +0.00%
BenchmarkTransportLogger/JSON-12                                 1928          1928          +0.00%
BenchmarkTransportLogger/JSON-Body-12                            6712          6712          +0.00%
BenchmarkBulkIndexer/Basic-12                                    178           178           +0.00%
BenchmarkJSONReader/None-12                                      576           576           +0.00%
BenchmarkJSONReader/Default-12                                   2769          2769          +0.00%
BenchmarkJSONReader/Default-Copy-12                              640           640           +0.00%
BenchmarkJSONReader/Custom-12                                    2275          2275          +0.00%
```